Proper Implementation of Json Web Tokens.
Now the server issues a JWT as soon as user logs in and sets them
in a HttpOnly cookie which will be sent at by the client to the server
in each subsequent requests.

The server sets two tokens in the cookie.
1) Access Token : This will be used to authenticate the user requests.
If the access token is valid, the request will be process or else Refresh
Token will be used to create a new access token.

2) Refresh Token : This token has a longer expiry time as compared to Access Token.
If access token is expired, a new token will be generated if the refresh token is valid.
In case both tokens expires, the user will need to login to the application again, with their credentials.

Additional features.
If the user email is not verified the user will able to login and explore the site.
To book a ticket. They need to verify their email.
Ticket booking functionality is not yet implemented.