Implemented the scanner for venue.
In the ventue, the supervisor or someone senior have to put the accessKey for the
location and the scanner will start.
Upon scanning the QR code, received on the email, all the booking details will be
displayed, which can be verified by the guards and entry will be provided.

Also fixed some bugs and changed the response messages of some requests.